requirements.txt
gwcomm-genwch
gwpd-genwch
python-dotenv
flask_restful
flask_jwt_extended
flask_restplus